requirements.txt
flask
flask_cors
opencv-python
numpy
pillow
pandas
scikit-learn
albumentations
ipykernel
python-dotenv